requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.8.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
h11==0.14.0
hf-xet==1.1.9
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
lxml==6.0.1
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==1.26.4
open-clip-torch==2.24.0
openai==1.107.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
pillow_heif==1.1.0
propcache==0.3.2
protobuf==6.32.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.16.2
sentencepiece==0.2.1
six==1.16.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.9.9
timm==1.0.19
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
transliterate==1.10.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xlrd==2.0.2
yarl==1.20.1
